rename transmute2 to transmute
