+++ src/closures/hello_closure.rs
@@ -13,7 +13,7 @@
 //
 
 use crate::{
-    transmute2::{LoanedCTypeRef, OwnedCTypeRef},
+    transmute::{LoanedCTypeRef, OwnedCTypeRef},
     z_loaned_hello_t,
 };
 use libc::c_void;
