+++ src/closures/matching_status_closure.rs
@@ -12,7 +12,7 @@
 //   ZettaScale Zenoh team, <zenoh@zettascale.tech>
 //
 use crate::{
-    transmute2::{LoanedCTypeRef, OwnedCTypeRef},
+    transmute::{LoanedCTypeRef, OwnedCTypeRef},
     zcu_matching_status_t,
 };
 use libc::c_void;
