+++ src/closures/query_channel.rs
@@ -13,7 +13,7 @@
 //
 
 use crate::{
-    transmute2::{LoanedCTypeRef, RustTypeRef, RustTypeRefUninit},
+    transmute::{LoanedCTypeRef, RustTypeRef, RustTypeRefUninit},
     z_loaned_query_t, z_owned_closure_query_t, z_owned_query_t,
 };
 use libc::c_void;
