+++ src/closures/reply_closure.rs
@@ -13,7 +13,7 @@
 //
 
 use crate::{
-    transmute2::{LoanedCTypeRef, OwnedCTypeRef},
+    transmute::{LoanedCTypeRef, OwnedCTypeRef},
     z_loaned_reply_t,
 };
 use libc::c_void;
