+++ src/liveliness.rs
@@ -18,10 +18,10 @@ use zenoh::{
     prelude::SessionDeclarations,
 };
 
-use crate::transmute2::LoanedCTypeRef;
+use crate::transmute::LoanedCTypeRef;
 use crate::{
     errors,
-    transmute2::{RustTypeRef, RustTypeRefUninit},
+    transmute::{RustTypeRef, RustTypeRefUninit},
     z_closure_reply_call, z_closure_sample_call, z_loaned_keyexpr_t, z_loaned_session_t,
     z_owned_closure_reply_t, z_owned_closure_sample_t, z_owned_subscriber_t,
 };
